backend/requirements.txt
django==4.2
djangorestframework==3.14.0
pymongo==4.7.2
python-dotenv==1.0.0

